refactor: remove AES scaffolding and narrow CLI scope
