+++ gnoman/commands/__init__.py
@@ -1,16 +1,11 @@
 """CLI subcommand handlers for GNOMAN."""
 
-from . import audit, autopilot, graph, guard, plugin, rescue, safe, secrets, sync, tx  # noqa: F401
+from . import audit, recover, safe, secrets, wallet  # noqa: F401
 
 __all__ = [
     "audit",
-    "autopilot",
-    "graph",
-    "guard",
-    "plugin",
-    "rescue",
+    "recover",
     "safe",
     "secrets",
-    "sync",
-    "tx",
+    "wallet",
 ]
