+++ gnoman/services/__init__.py
@@ -1,0 +1,17 @@
+"""Runtime service accessors for GNOMAN."""
+
+from .state import (
+    get_audit_reporter,
+    get_audit_signer,
+    get_recovery_toolkit,
+    get_safe_vault,
+    get_secrets_store,
+)
+
+__all__ = [
+    "get_audit_reporter",
+    "get_audit_signer",
+    "get_recovery_toolkit",
+    "get_safe_vault",
+    "get_secrets_store",
+]
